requirements.txt
flask
flask-cors
flask-docs
flask-mail
flask-sqlalchemy
numpy==1.23.5
openai==1.47.0
openpyxl==3.1.2
python-docx==0.8.11
python-dotenv==1.0.1
pymysql
flasgger
tiktoken
markdown
langchain
pymupdf
frontend
langchain_openai
PyPDF2
faiss-cpu
pymdown-extensions
latex2mathml==3.77.0
rouge==1.0.1
nltk==3.9.1
# ra_check功能新增依赖
scikit-learn
tqdm
httpx==0.27.0